mildoc_zbb/mildoc_wxkf/requirements.txt
# 环境变量管理
python-dotenv==1.2.1

# 日期处理
pytz==2025.2

# Flask Web框架
Flask==3.1.2
Flask-Login==0.6.3

pycryptodome==3.23.0
requests==2.32.5

# PDF文档解析
PyPDF2==3.0.1

# Office文档解析
markitdown[all]==0.1.4

# Minio对象存储客户端
minio==7.2.20

# Milvus向量数据库客户端
pymilvus==2.6.5


# OpenAI API客户端（用于embedding）
openai==2.11.0

# langchain相关依赖
langchain==1.1.3
langchain-openai==1.1.3
langchain-milvus==0.3.1
langchain-community==0.4.1
langchain-core==1.2.0
langchain-text-splitters==1.1.0

# 阿里云
alibabacloud-oss-v2==1.2.1
aliyun-mns-sdk==1.3.0